+++ packages/core-api/lib/versions/2/methods/delegates.js
@@ -86,10 +86,12 @@ const voterBalances = async request => {
 }
 
 module.exports = server => {
+  const generateTimeout = require('../../utils').getCacheTimeout()
+
   server.method('v2.delegates.index', index, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -102,7 +104,7 @@ module.exports = server => {
   server.method('v2.delegates.show', show, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request => generateCacheKey({ id: request.params.id }),
@@ -111,7 +113,7 @@ module.exports = server => {
   server.method('v2.delegates.search', search, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -125,7 +127,7 @@ module.exports = server => {
   server.method('v2.delegates.blocks', blocks, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -138,7 +140,7 @@ module.exports = server => {
   server.method('v2.delegates.voters', voters, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -151,7 +153,7 @@ module.exports = server => {
   server.method('v2.delegates.voterBalances', voterBalances, {
     cache: {
       expiresIn: 8 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request => generateCacheKey({ id: request.params.id }),