+++ packages/core-api/lib/versions/2/methods/wallets.js
@@ -116,10 +116,12 @@ const search = async request => {
 }
 
 module.exports = server => {
+  const generateTimeout = require('../../utils').getCacheTimeout()
+
   server.method('v2.wallets.index', index, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -133,7 +135,7 @@ module.exports = server => {
   server.method('v2.wallets.top', top, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request => generateCacheKey(utils.paginate(request)),
@@ -142,7 +144,7 @@ module.exports = server => {
   server.method('v2.wallets.show', show, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request => generateCacheKey({ id: request.params.id }),
@@ -151,7 +153,7 @@ module.exports = server => {
   server.method('v2.wallets.transactions', transactions, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -165,7 +167,7 @@ module.exports = server => {
   server.method('v2.wallets.transactionsSent', transactionsSent, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -179,7 +181,7 @@ module.exports = server => {
   server.method('v2.wallets.transactionsReceived', transactionsReceived, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -193,7 +195,7 @@ module.exports = server => {
   server.method('v2.wallets.votes', votes, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>
@@ -206,7 +208,7 @@ module.exports = server => {
   server.method('v2.wallets.search', search, {
     cache: {
       expiresIn: 30 * 1000,
-      generateTimeout: 3000,
+      generateTimeout,
       getDecoratedValue: true,
     },
     generateKey: request =>