+++ packages/core-api/lib/versions/utils.js
@@ -1,0 +1,7 @@
+exports.getCacheTimeout = () => {
+  const {
+    generateTimeout,
+  } = require('@arkecosystem/core-container').resolveOptions('api').cache
+
+  return JSON.parse(generateTimeout)
+}